requirements.txt
aniso8601==9.0.1
anyio==3.6.2
aplus==0.11.0
astropy==5.2.1
asttokens==2.2.1
autopep8==2.0.1
backcall==0.2.0
blake3==0.3.3
bqplot==0.12.36
branca==0.6.0
cachetools==5.2.1
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
cloudpickle==2.2.0
colorama==0.4.6
comm==0.1.2
commonmark==0.9.1
contourpy==1.0.6
coverage==7.0.4
cycler==0.11.0
dask==2022.12.1
debugpy==1.6.5
decorator==5.1.1
entrypoints==0.4
executing==1.2.0
fastapi==0.89.0
filelock==3.9.0
Flask==2.2.2
Flask-Cors==3.0.10
Flask-JSON==0.3.5
Flask-RESTful==0.3.9
fonttools==4.38.0
frozendict==2.3.4
fsspec==2022.11.0
future==0.18.2
h11==0.14.0
h5py==3.7.0
httptools==0.5.0
idna==3.4
importlib-metadata==6.0.0
ipydatawidgets==4.3.2
ipykernel==6.20.1
ipyleaflet==0.17.2
ipympl==0.9.2
ipython==8.8.0
ipython-genutils==0.2.0
ipyvolume==0.5.2
ipyvue==1.8.0
ipyvuetify==1.8.4
ipywebrtc==0.6.0
ipywidgets==8.0.4
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jupyter_client==7.4.8
jupyter_core==5.1.3
jupyterlab-widgets==3.0.5
kiwisolver==1.4.4
llvmlite==0.39.1
locket==1.0.0
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numba==0.56.4
numpy==1.23.5
packaging==23.0
pandas==1.5.2
pandas-highcharts==0.5.2
parso==0.8.3
partd==1.3.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==2.6.2
progressbar2==4.2.0
prompt-toolkit==3.0.36
psutil==5.9.4
pure-eval==0.2.2
pyarrow==10.0.1
pycodestyle==2.10.0
pydantic==1.10.4
pyerfa==2.0.0.1
Pygments==2.14.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
python-utils==3.4.5
pythreejs==2.4.1
pytz==2022.7
pywin32==305
PyYAML==6.0
pyzmq==24.0.1
requests==2.28.1
rich==13.0.1
six==1.16.0
sniffio==1.3.0
stack-data==0.6.2
starlette==0.22.0
tabulate==0.9.0
tomli==2.0.1
toolz==0.12.0
tornado==6.2
traitlets==5.8.0
traittypes==0.2.1
typing_extensions==4.4.0
urllib3==1.26.13
uvicorn==0.20.0
vaex==4.16.0
vaex-astro==0.9.3
vaex-core==4.16.1
vaex-hdf5==0.14.1
vaex-jupyter==0.8.1
vaex-ml==0.18.1
vaex-server==0.8.1
vaex-viz==0.5.4
watchfiles==0.18.1
wcwidth==0.2.5
websockets==10.4
Werkzeug==2.2.2
widgetsnbextension==4.0.5
xarray==2022.12.0
xyzservices==2022.9.0
zipp==3.11.0
